hub/xinet/requirements.txt
fastapi
plotly
uvicorn
jupyter-dash
